requirements.txt
# Requirements for Dungeon Generator
numpy
networkx
Pillow
jsonschema
pytest
setuptools
colour
noise
pylint
colorlog
